pyproject.toml
[project]
name = "watt-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
    "azure-identity>=1.25.2",
    "azure-keyvault-secrets>=4.10.0",
    "azure-storage-file-datalake>=12.23.0",
    "parsel>=1.11.0",
    "pandas>=2.0.0",
    "requests>=2.32.5",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
]

[dependency-groups]
dev = [
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pyodbc>=5.2.0",
    "python-docx>=1.2.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
